requirements-dev.txt
django-extensions==2.1.0
markdown==2.6.11
requests-mock==1.5.2
